benchmarks/framework-shootout/apps/flask_app/requirements.txt
flask==3.1.*
gunicorn==23.*
psycopg2-binary==2.9.*
orjson==3.*
